Renamed packages - renamed inf112.skeleton.app to inf112.roborally.game - renamed GameWorld to world - renamed GameObjects to objects
- also smaller changes in MovableGameObjects class and in Direction enum
