+++ src/main/java/inf112/roborally/game/objects/Flag.java
@@ -1,4 +1,4 @@
-package inf112.skeleton.app.GameObjects;
+package inf112.roborally.game.objects;
 
 public class Flag {
     private int flagNumber;
