+++ src/main/java/inf112/roborally/game/objects/Rotate.java
@@ -1,4 +1,4 @@
-package inf112.skeleton.app;
+package inf112.roborally.game.objects;
 
 
 public enum Rotate {
