+++ src/test/java/inf112/roborally/game/AppTest.java
@@ -1,4 +1,4 @@
-package inf112.skeleton.app;
+package inf112.roborally.game;
 
 import static org.junit.Assert.assertTrue;
 import org.junit.Test;
