+++ src/test/java/inf112/roborally/game/ProgramCardTest.java
@@ -1,5 +1,7 @@
-package inf112.skeleton.app;
+package inf112.roborally.game;
 
+import inf112.roborally.game.ProgramCard;
+import inf112.roborally.game.objects.Rotate;
 import org.junit.Before;
 import org.junit.Test;
 import java.util.Stack;
